Store messages on disk instead of per process
This allows a client and a daemon to share messages, without having to run a
sync. Currently, there is no locking, but as files are never removed or
modified, there shouldn't be a need for locking.

Files are stored using the same format that they are sent across the network,
using .serialise and .from_serialised