Split client.py
This allows other programs to use client.py as a base to implement their own
clients. cli.py is the exact same interface as what was in client.py, but
implemented as a different file, and using client.py as a library.

This will make it easier to add a daemon which runs the sync operation on a
list of known hosts, without having to copy paste the sync() code.