Move network functionality into util.py
This removes the duplicate code that was present in both the daemon and the
client, and will make it easier to add new functionality later on.

This has a high chance of causing bugs, and has been tested for basic syncing
to a different machine, but not very well. However, in the long run, this will
make the code less buggy, as any changes only need to be made once.