Initial work towards ignoring keys.
As of right now, the ignored keys are not mentioned, and will be treated as if
they simply don't exist. This will waste disk space and bandwidth if one client
has a key ignored but the other one doesn't. I will modify the protocol to
include a list of keys that would be sent, but shouldn't due to the client or
server ignoring them.