i hate network programming
Anyway, this SHOULD fix the invalid MSGID bug. The problem was that I was
reading parts of the next message, because I was checking the length every 64
bytes. Doing a recv(1) would fix this as well, but I think this is a cleaner
solution.

This has only been tested on localhost, however. I will push the changes to
master when I've tested it across the LAN and it works fine.