Allow running commands as arguments to cli.py
The first argument will be taken as the command, with the next ones being
passed as arguments to the command.

Since commands have a way to indicate failure (returning 1 or returning
nothing, which is seen as NoneType), the exit code of cli.py will be the exit
code of the command, which makes it easier to detect failure in scripts.