Add function to calculate cost
Sending a message has a base cost of ~ 1 second (measured on my machine)

The way costs are done is that since it takes 1 second to calculate 1 million
hashes, if you want a delay of 1 second the hash, converted to an integer,
should be lower than "2**256 / 1 million". This allows for more precise delays
than just using amount of bits required, but makes it more complex.

This formula may be tweaked, but changes may mean that messages are dropped
from older clients, so it shouldn't be changed very often.