Very basic support for custom line editor
Doing this partly for the practice, and partly because readline is GPL.
Reminds me, I should get around to putting a license on this.

It only supports very basic line editing tools, such as left/right arrow keys,
backspace, and typing in the middle of a string. If it crashes the terminal
will be stuck in raw mode, so you'll have to type reset.